+++ src/test/testSprintf.c
@@ -29,7 +29,7 @@ static void testBasicSprintf(MprTestGroup *gp)
     assert(strcmp(buf, "-12345678") == 0);
 
     str = mprAsprintf(gp, 4, "%d", 12345678);
-    count = strlen(str);
+    count = (int) strlen(str);
     assert(count == 3);
     assert(strcmp(str, "123") == 0);
     mprFree(str);
